{{cookiecutter.project.slug}}/requirements/base.txt
Django<2.0
django-crispy-forms==1.6.1
django-environ==0.4.4
django-smartfields==1.0.9
mysqlclient
Pillow==4.3.0
psycopg2
pytz==2017.2
six==1.11.0
sorl-thumbnail==12.3
django-cms<3.5
djangocms-cascade==0.15.4
django-filer==1.3.0
django-mptt==0.8.7
django-polymorphic==1.3
easy-thumbnails==2.4.2
djangocms-attributes-field==0.3.0
djangocms-video==2.0.4
djangocms-snippet==1.9.2
